mf0env/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.2.0
pytz==2022.5
sqlparse==0.4.2
tzdata==2022.2
